update xuathang & changeLongToInt mysql

+++ backend/src/main/java/com/example/warehouse/controller/ExportReceiptDetailController.java
@@ -23,7 +23,7 @@ public class ExportReceiptDetailController {
     }
 
     @GetMapping("/{maPhieuXuat}/{maSanPham}")
-    public ResponseEntity<ExportReceiptDetail> getDetailById(@PathVariable Long maPhieuXuat,
+    public ResponseEntity<ExportReceiptDetail> getDetailById(@PathVariable int maPhieuXuat,
             @PathVariable String maSanPham) {
         ExportReceiptDetailId id = new ExportReceiptDetailId();
         id.setMaPhieuXuat(maPhieuXuat);
@@ -47,7 +47,7 @@ public class ExportReceiptDetailController {
     }
 
     @PutMapping("/{maPhieuXuat}/{maSanPham}")
-    public ResponseEntity<ExportReceiptDetail> updateDetail(@PathVariable Long maPhieuXuat,
+    public ResponseEntity<ExportReceiptDetail> updateDetail(@PathVariable int maPhieuXuat,
             @PathVariable String maSanPham,
             @RequestBody ExportReceiptDetail detail) {
         ExportReceiptDetailId id = new ExportReceiptDetailId();
@@ -62,7 +62,7 @@ public class ExportReceiptDetailController {
     }
 
     @DeleteMapping("/{maPhieuXuat}/{maSanPham}")
-    public ResponseEntity<Void> deleteDetail(@PathVariable Long maPhieuXuat,
+    public ResponseEntity<Void> deleteDetail(@PathVariable int maPhieuXuat,
             @PathVariable String maSanPham) {
         ExportReceiptDetailId id = new ExportReceiptDetailId();
         id.setMaPhieuXuat(maPhieuXuat);
