+++ backend/src/main/java/com/example/warehouse/model/ExportReceipt.java
@@ -19,7 +19,7 @@ public class ExportReceipt {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ma_phieu_xuat")
     @JsonProperty("maPhieuXuat")
-    private Long maPhieuXuat;
+    private int maPhieuXuat;
 
     @Column(name = "ngay_xuat")
     @JsonProperty("ngayXuat")
